Switch ImageDecoder over to new ImageTrack semantics.
This moves frameCount and repetitionCount onto the track properties
and removes selectTrack() in favor of "track.selected = true". The
tracks list is modeled after Audio/VideoTrackList, but for now does
not include any events.

This also hardens the 'animated' status on a track so that it won't
change after metadata has been retrieved. repetitionCount is switched
to a 'unrestricted float' so 'Infinity' can be be accurately signaled.

R=chcunningham

Bug: 1182435, 1183456
Test: New and updated unittests and wpt tests.
Change-Id: Ibf8363e6f80d998202d42f745180812953681da9