Refactor 'last baseline' tests out of grid-abspos-staticpos
Neither Blink nor Webkit supports the 'last baseline' keyword. In order
to get a better sense of interop as part of the #compat2021 effort,
this change refactors out the portion of the css-grid/abspos tests that
use last baseline.

There are both legacy and NG failures in the remaining cases in the
tests, which are tracked by new bugs.

Change-Id: Ifdcfd2619b83ed0e13a926ab832a9d58e38688ae